Skip RPM tests if RPM plugin not installed
If a given plugin is not installed on Pulp, none of the tests for that
plugin should run. Module
`pulp_smash.tests.rpm.api_v2.test_packages_signature` is missing that
skipping logic. Add it. (Checks on a per-module basis like this would be
unnecessary if the `load_modules` unittest protocol were in use.)