Refactor just-added RPM rsync distributor tests
Add a new function, `_get_dists_by_type_id`. It finds information about
all of a repository's distributors, and returns that information as a
dict keyed by distributor type IDs. The dict is in this form:

    {'yum_distributor': {…}, 'rpm_rsync_distributor': {…}}

Add a new method, `_RsyncDistUtilsMixin.make_repo`. It creates a
repository with a yum importer, yum distributor and RPM rsync
distributor.

Finally, rename `_MakeUserMixin` to `_RsyncDistUtilsMixin`. This
reflects the fact that the class contains tools for working with RPM
rsync distributors, not just users.

These changes reduce the amount of duplicated logic in the test cases
(`PublishTestCase` and `PublishBeforeYumTestCase`), and should be useful
when writing more test cases.

The RPM rsync distributor test cases were added in
96fc3884b070965edfe00bb05eceb58f0305830d.