Update assertion for PublishBeforeYumDistTestCase
`PublishBeforeYumDistTestCase` creates an RPM repository with both a yum
distributor and an RPM rsync distributor, and it publishes with the
latter before the former. Currently, the test case asserts that the
publish fails. However, according to Pulp issue 2187, this should result
in the publish skipping. Update the test case accordingly.

See: https://pulp.plan.io/issues/2187