Expand test for pulp-admin's --force-sync option
Expand `pulp_smash.tests.rpm.cli.test_sync.ForceSyncTestCase`.
Currently, the test case will:

1. Create and sync a repo.
2. Remove an RPM. Verify one RPM is missing.
3. Sync the repo with `--force-full true`. Verify all RPMs are present.

This test procedure shows that the second sync restores the removed RPM,
but it doesn't show that the `--force-sync true` option was the cause of
that restoration. It could be that syncs *always* restore removed RPMs.
(This would indicate a different issue with the fast-forward logic.)
Update the test so that it will:

1. Create and sync a repo.
2. Remove an RPM. Verify one RPM is missing.
3. Sync the repo. Verify one RPM is missing.
4. Sync the repo with `--force-full true`. Verify all RPMs are present.