Use orphans to speed up RPM rsync distrib tests
If a sync is performed and matching orphans are on disk, Pulp will
re-use them instead of syncing downloading the content units from a
remote system.

Update the test cases for the RPM rsync distributor. Drop the call to
delete orphans after each test case, and instead delete orphans when the
module has finished executing.