Drop constant RPM_ABS_PATH
The `RPM_ABS_PATH` constant defines the absolute path to the `RPM`
content unit on the Pulp filesystem.

For this constant to hold true, the `RPM` content unit must not change.
However, it has changed at least twice during the history of Pulp
Fixtures, as signatures have been added and removed. Also, its signature
might change in the future if the public/private keypair in Pulp
Fixtures changes for some reason, and the exact path to the content unit
is supposed to be an implementation detail.

Rewrite the one test case that uses this constant to use the `find` CLI
executable instead.