Fix test for RPM rsync distrib's "delete" option
When publishing with an RPM rsync distributor, one can pass a "delete"
option. This option specifies that files which don't exist in the RPM
repository on the source system should be deleted from the RPM
repository on the remote system.

The path to the RPMs on the remote system is constructed from values in
both the yum distributor and the RPM rsync distributor. Module
`pulp_smash.tests.rpm.api_v2.test_rsync_distributor` fails to implement
correct path-building logic. Fix this error, and expand the module-wide
documentation, so that this error will be easier to avoid in the future.

Fix some other minor issues. Drop a duplicate `skipTest` call. Fix an
assertion.