Make test case more resilient against errors
Add a new method, `pulp_smash.selectors.skip_if`. This method is similar
to `@unittest.skipIf`, but more flexible: it lets one execute skipping
logic at runtime and is capable of accessing class and instance
attributes.

Why make this method available? It exists because of the structure of
the unittest module:

1. The `setUpClass` method lacks access to certain tools like the
   `addCleanup` method. As a result, it is desirable to move
   failure-prone code to other locations, such as test methods.
2. Some failure-prone code should control whether or not the test
   methods in a test case should execute, but this is difficult to do,
   because all test methods execute regardless of how many test methods
   execute.
3. The `skip_if` decorator lets one move failure-prone code into a
   regular test method *and* control whether subsequent test methods
   execute.

This decorator may make it easier to avoid
`pulp_smash.utils.BaseAPITestCase`.