Replace Service with ServiceManager
The `Service` class represents a service (such as httpd) on a host. It
provides methods such as `start` and `stop`, and it abstracts away which
service manager is in use on the target host.

A drawback of the `Service` class is that it encourages serial code to
be written. If three services are to be restarted, it's easy to write
something like this:

    services = (Service(cfg, name) for name in ('foo', 'bar', 'biz'))
    for service in services:
        service.stop()
    for service in services:
        service.start()

The `ServiceManager` class represents the service manager (such as
systemctl) on a host. It's used like so:

    services = ('foo', 'bar', 'biz')
    svc_mgr = ServiceManager(cfg)
    svc_mgr.stop(services)
    svc_mgr.start(services)

Depending on the service manager used by the target system, `stop` and
`start` dispatch commands to the target system either in serial or in
parallel. This has the potential to improve runtime performance.