Fix how command is built
Module `pulp_smash.tests.platform.cli.test_selinux` contains CLI
command-building logic. Here's an example of a command that might be
built:

    getfattr --recursive --name=security.selinux /var/cache/pulp

Unfortunately, if `sudo` is required, the following command is built:

    sudo --recursive getfattr --name=security.selinux /var/cache/pulp

Fix this. Build the CLI command in strictly left-to-right order, in
order to ease human understanding of what's going on.