Change how package paths are calculated
The paths to packages in published repositories have changed. In Pulp
2.11 and earlier, packages ended up in the root of published
repositories, alongside the "repodata" directory. In Pulp 2.12, packages
default to being placed elsewhere, and users have the option to
customize the exact path.

The end result of this change is that clients should consult the
`repodata/...primary.xml` file when finding packages. This commit is not
so elegant, and instead implements a stop-gap solution. This stop-gap
solution should fix many broken tests, but does not yet consult with the
XML files.

See: https://github.com/PulpQE/pulp-smash/issues/472