Skip certain Python tests
Version 2 of the pulp_python plugin will add support for dealing with
python wheels, and most of the necessary work is already done. As a
result, certain issues on Pulp's issue tracker are marked as MODIFIED,
and several of Pulp Smash's selectors designed to skip tests were not
producing test skips. Fix this.