Disable SELinux when managing services on Jenkins
When managing SYSV services from Jenkins services and related files gets
the `unconfined_java_t` label. That is because SYSV does not deal great
with the labels as systemd. In order to get better results, on RHEL6,
when running on Jenkins disable SELinux temporarily when restarting the
services for pulp_reset.