Use Puppet fixtures from Pulp Fixtures
Pulp Fixtures has grown the ability to generate a Puppet module. Make
use of this module, where possible.

The new fixtures provided by Pulp Fixtures are not comprehensive, so
there are still references to the Puppet Forge. To make this distinction
clear, and to ensure that users default to using the fixtures provided
by Puppet Fixtures when possible:

* Append ``_1`` to fixtures provided by Pulp Fixtures.
* Append ``_2`` to fixtures provided by the Puppet Forge.