Fix some broken errata tests
Historically, Pulp has always published errata in an RPM repository,
regardless of whether those errata reference any packages, and
regardless of whether referenced packages exist in the same published
RPM repository.

The current development versions of Pulp 2.12 and 2.13 change how errata
are handled. Now, an errata is only included in a published RPM
repository if it references a package, and if that package is present in
the same published repository.

Update `UpdateInfoTestCase`, so that it observes this new behaviour.

See: https://pulp.plan.io/issues/2559