Rewrite test that updates updateinfo.xml
Class `ErratumPkgListCountTestCase` does the following:

1. Create and sync an RPM repository with a feed and distributor.
2. Unassociate the "gorilla" package from the repository.
3. Publish the repository.

It then verifies that the published repository's `updateinfo.xml` file
is valid. Unfortunately, this test recently broke, primarily because of
the changes implemented for Pulp issue 2559: errata are omitted if they
don't reference any packages, or if they reference packages not in the
currently published repository. In this case, an erratum that references
the "gorilla" erratum is now being omitted, which broke the test.

Update the test case:

* Make the test case comply with the changes outlined in Pulp issue
  2559.
* Make the test clean up after itself even when an error occurs during
  test case set-up.
* Rename the test case to `UpdateRepoTestCase`, and modify its logic.

The test case now does the following:

1. Create, sync and publish an RPM repository with a feed and
   distributor.
2. Verify the repository's `updateinfo.xml` file.
3. Unassociate the "gorilla" package from the repository.
4. Verify the repository's `updateinfo.xml` file.

See:

* https://github.com/PulpQE/pulp-smash/issues/540
* https://github.com/PulpQE/pulp-smash/pull/574
* https://pulp.plan.io/issues/2559