Update tests for `updateinfo.xml`
When Pulp syncs and publishes a repository, it's free to restructure the
repository at will, so long as the result is semantically valid. One
such restructuring is to condense multiple `<pkglist>` elements. Another
is to change the "short" attribute of a `<collection>` element. There
are more.

Update module `pulp_smash.tests.rpm.api_v2.test_updateinfo` so that its
test cases only make assertions on important information, like errata
IDs and the presence of certain packages.