Refactor OSTree distributor test cases
An OSTree distributor may have a `relative_path` defined. Relative paths
must be globally unique within a Pulp installation. The following test
cases verify that this is so:

* `CreateDistributorsTestCase`
* `UpdateDistributorsTestCase`

Refactor the test cases so that clean-up is more deterministic, and so
that test failures are easier to diagnose.

Also, skip tests affected by Pulp issue 2769.

See: https://pulp.plan.io/issues/2769