Set Docker upstream name based on Pulp version
Pulp 2.14 gained the ability to sync repositories with manifest lists.
To test this functionality, one must sync content from upstream
repositories that have manifest lists. Unfortunately, when older
versions of Pulp encounter an upstream repository with a manifest list,
they just break.

Let there be two constants for Docker upstream names:

* `DOCKER_UPSTREAM_NAME`
* `DOCKER_UPSTREAM_NAME_NOLIST`

Add a function that returns the appropriate constant (in most cases)
based on the version of Pulp under test, and use it.