Fix make targets for publishing Pulp Smash
The "package" make target is phony, meaning that it always triggers
packaging logic, even if packages already exist. This is problematic: if
one runs the "package" make target twice in a row, it will produce brand
new packages.

While this may seem like a good thing, it's actually problematic. If one
calls `scripts/release.sh` and the resultant package(s) are verified as
being good quality, then the "publish" make target shouldn't overwrite
those verified-good packages. It's true that the packaging process
should produce identical packages every time. But why overwrite
something that's known to be good, and for no good reason?

Rename the "package" make target to "dist," and make it a regular make
target instead of a phony one. As a result of this change, the packaging
make target no longer overwrites existing packages.