Test actions over a non-existent repository using API.
Atemmpt to perform the following actions over a non-existent
repository: Sync, Publish, Update and Delete.

In all these actions is expected an exception to be raised. HTTP 4XX.

These actions are testing the failures scenarios mentioned in the
issue #157.

Refactor of `RPMDistributorTestCase` to remove `BaseAPITestCase` as
base class, and remove `config` variable.

Closes: #157