Ignore flake8 warnings about bare keywords
It's dangerous to catch an exception with a bare `except` keyword.
However, there are legitimate use cases doing so. Update the code so
that the existing bare catches are allowed.

There's at least two ways to allow the use of bare `except` keywords:
ignore E722 globally, or ignore E722 on a per-line-of-code basis. Do the
latter, on the theory that bare excepts *are* dangerous, and the user
should be forced to explicitly acknowledge the danger in what they're
doing.