Fix how raw command is executed, for MtimeTestCase
Bash interprets commands in the following order:

1. Lexical analysis and parsing
2. Macro expansion
3. Command execution

This has some interesting implications. Let's say the following files
exist:

* `/root/bar`
* `/root/foo`

If root executes `ls /root/*`, then the following will be printed:

    bar  foo

But if a regular user executes that same command, it will fail, with an
error stating that `/root/*` doesn't exist. This is because regular
users can't see those files, so the star isn't expanded during the macro
expansion step.

One way to bring some consistency to this situation is to execute a
command like `bash -c 'ls /root/*'`. Use this strategy to fix a failure
in `MtimeTestCase`.