Drop redundant function for getting base URL
The following functions perform essentially the same task:

* `pulp_smash.config.PulpSmashConfig.base_url`
* `pulp_smash.config.PulpSmashConfig.get_base_url`

The difference is that `base_url` is a property that can't accept
arguments, whereas `get_base_url` is a function that can accept
functions. Drop the former in favor of the latter. The `base_url`
property was useful some time back during a transition period when the
config file format changed.