Cache JWT authentication tokens
Make the `get_auth` function return a cached JWT authentication object
when possible. This makes it possible to call `get_auth` with abandon,
without worrying about generating a huge number of tokens.

The cache used by the function can be easily cleared. This will likely
be useful in the future, when there is code that invalidates JWT tokens,
or that resets Pulp, or so on.