Let PulpSmashConfig.read traverse alternate paths
By default, this method reads `~/.config/pulp_smash/settings.json`. But
it's designed to be flexible enough to deal with other directory or file
names. It already respects the `PULP_SMASH_CONFIG_FILE` environment
variable. But it should also respect arguments passed directly to the
read() function.

As a result of this change, there are three sources of information that
can affect read's behaviour. In ascending priority:

1. Built-in defaults.
2. The `PULP_SMASH_CONFIG_FILE` environment variable.
3. Arguments to read().