Make lint-pylint make target faster and stricter
The lint-pylint make target runs pylint over nearly the entire code
base. The make target calls pylint twice, to check different parts of
the code base. The code in `pulp_smash/tests/` is not subject to the
duplicate-code check, and the remaining code is.

This distinction was implemented when Pulp Smash was a much newer
entity. At that time, the goal was to build up a functional testing
suite quickly, and loosening checks on the functional tests was
considered a reasonable shortcut to achieve that goal faster. Since
then, the code base has matured, and strategies for dealing with
duplicate code have been implemented. The `utils` modules are a good
example.

There are two down-sides to allowing this distinction to live on:

* A bad coding practice is given a pass within certain parts of the code
  base.
* The fact that the `lint-pylint` job is implemented as a pair of tasks
  means that a serial bottleneck is introduced into the job, the impact
  of which should be especially pronounced on modern multi-core CPUs.
  [1]

Erase the distinction between the two sections of the code base.

Some quick benchmarks show that:

* On a host with a 4C/8T CPU, lint-pylint execution time drops by 8%.
* On a host with an 8C/16T CPU, lint-pylint execution time drops by 20%.