Test each combo of download policy and sync mode
Expand the recently added `SyncFileRepoTestCase`. Currently, this test
case creates an importer with a random download policy and a random sync
mode. With this commit, the test case will instead iterate through each
possible pairing of download policy and sync mode.

This change in test logic uncovered a bug, wherein not all of the
download policies are implemented for the file plugin.

See: https://pulp.plan.io/issues/3320