Test upload of a manifest list V2 without a sync.
Move `SyncPublishMixin` to docker/api_v2/utils. This class will be used
as a helper to existent tests, and for new ones. This change is
necessary to avoid DRY.

Refactor and rename existent tests to reflect the previous change.

Create a helper function `write_manifest_list` on docker/utils.

Do the following:

 1. Create, sync and publish a repository.
 2. Read the manifest list,and remove one of the existent architectures
  from it.
 3. From the modified manifest list create a new JSON valid on a
  temporary dir on disk.
 4. Upload the just created JSON file to the repository.
 5. Assert that the number of `docker_manifest_list` present on the
  repository has increased.

Closes: #829