Add a template for Pulp-Smash issues.
In order to create a standard, and make easy to a new contributor to
file an issue it was included a template with the the basics steps in
how to file a pulp-smash issue.

It was added a dir `.github` to the root system of the project, thus
avoiding to add extra files to the root system of smash.

Futures templates can be added to the same dir as well.