Add tests for CRD artifacts.
As of this writing, pulp3 does not handle upload of duplicate files.
Due that, there is a need to clean the artifacts present in pulp3, before
attempt to upload a new file to pulp3.
However, if a file exists as content unit and as artifact, the file has
to be deleted first as a content unit then as an artifact.
To solve this issue, 2 new functions were added to pulp3/utils.

- `clean_content_units()`
- `clean_artifacts()`

Add tests to CRD artifacts.

1. Create an artifact by uploading a file to pulp3.
2. Reads an artifact by its href.
3. Delete an artifact.

Closes: #726