Add tests to CRUD content unit.
In order to create a content unit at least one artifact has to exist in
Pulp3, then metadata related to this artifact will be used to create a content
unit.

An artifact is just data on disk, a content unit is an entry in a
database that points to an artifact and gives it a name.

Closes:#870