Drop warning about base_path trailing slash
`pulp_smash.tests.pulp3.pulpcore.utils.gen_distribution()` returns a
dict that can be used to generate a distribution. A `base_path` is a
required attribute for a distribution, and as such, this function must
answer the following question:

> May a distribution's `base_path` have a trailing slash?

Pulp issue 3412 asks this question more formally. It has been closed,
and the answer is "no, a distribution's `base_path` may not have a
trailing slash."

Drop a warning about this issue from `gen_distribution()`.