Refactor just-added test for repository versions
Do the following:

* Split two logically separate portions of code into two separate test
  methods.
* Make clean-up logic more robust.
* Expand docstrings.
* Reformat some bits of code.

Notably, this refactor doesn't completely overhaul the test structure or
expand the number of assertions performed. The goal is to take something
that works and make it more robust. A future commit may try to expand
the number of assertions, or perform more drastic flow control
alterations.