Get fresh repo data
Many of the utility functions in `pulp_smash.tests.pulp3.utils` were
recently re-written to be faster and simpler. The flip-side of this
change is that they're prone to making use of out-of-date information.
The test case in
`pulp_smash.tests.pulp3.file/api_v3.test_download_content` is making use
of one of these utility functions, and has not been brought up-to-date
with the design change.

See: 3907dce380054b1a37cf17a7a4b85cb4a078210c