SELinux tests need to a toggle for some test environments
The selinux tests currently require the vm/container to be running
selinux. This causes an issue on test environments such as OX +
Containers which cannot support selinux.

So we now provide a settings.json to be able to let the user/provisoner
toggle this setting off. The default behavior will of course be to
require (and expect) selinux tests to pass. We looked at trying to
detect if selinux is disabled, but it is possible for a bug elsewhere to
disable tests we want to always run. So flag, instead.