Replace get_version_hrefs() with get_versions()
High-level general-purpose utilities are most useful when they return
complete or near-complete responses from Pulp, instead of attributes
extracted from Pulp's responses. The `get_version_hrefs` function
violates this design principle: it fetches the versions of a repository,
and then returns just the hrefs of each version.

Drop this function in favor of a `get_versions` function, which returns
the complete list of repository versions returned by Pulp. This lets the
function be used in more places, such as in the `get_repo_versions_attr`
function.

In addition, let the revamped function accept a `params` argument, which
can be used to filter which repositories Pulp returns. This lets it
replace the `filter_repo_version` function.