Make final argument to PulpSmashConfig kwarg only
The PulpSmashConfig constructor may gain additional arguments in the
future. For example, it might gain a `pulp_fips_enabled` argument,
stating whether FIPS tests should be enabled. If this occurs, it would
be ideal if this argument could be added after the other `pulp_*`
arguments, all of which define system-wide information, and before the
`hosts` argument, which defines per-host information. This can only
seamlessly occur if all users of the `PulpSmashConfig` object pass the
final `hosts` argument as a keyword argument, instead of as a positional
argument.

Make the final `hosts` argument to the `PulpSmashConfig` constructor
keyword-only.