Decouple @require and @skip_if from unittest
The `@require` decorator raises a `unittest.SkipTest` exception if the
Pulp application under test appears to be too old. Similarly, the
`@skip_if` decorator raises a `unittest.SkipIf` exception if a certain
test evaluates to true. These behaviours are great when the test runner
being used is unittest, or something largely compatible with unittest.
But Pulp Smash should be compatible with any test runner, not just
unittest.

Add a new argument to both `@require` and `@skip_if`, where the new
argument is the name of an exception to be raised. Let it default to
`unittest.SkipTest`, which should be fine in most cases.

See: https://github.com/PulpQE/pulp-smash/issues/1033