Decouple Pulp 3 utils from unittest
The `@require_pulp_3` and `@require_pulp_plugins` Pulp 3 decorators
raise a `unittest.SkipTest` exception under certain circumstances. While
this functionality is fantastically useful, it's tightly coupled to
unittest. This is problematic, given that Pulp Smash should be
compatible with any test runner, not just unittest.

Add a new argument to these two decorators, where the new argument is
the name of an exception to be raised. Let it default to
`unittest.SkipTest`, which should be fine in most cases.

See: https://github.com/PulpQE/pulp-smash/issues/1033