Drop unittest from Pulp 2 require_* utils
The various `require_*` methods in `pulp_smash.pulp2.utils` are designed
to skip tests under certain circumstances. This is easiest to do if some
assumptions are made about which test runner is being used. For example,
if one assumes that a unittest-compatible test runner is likely to be
used, then it makes sense for the functions to raise
`unittest.SkipTest`.

However, Pulp Smash is being turned into a library that should be
compatible with other test runners. An effective way to help prevent
bias toward unittest is to avoid any references to unittest within the
code base. Add an `exc` parameter to the functions, which is the
exception to be instantiated and raised.

See: https://github.com/PulpQE/pulp-smash/issues/1033