Verify erratum are listed in applicability reports
When Pulp returns a content applicability report for a consumer, it
should list not only the RPMs that are applicable, but also the erratum.
Pulp 2.16.2 included a regression wherein this information would not be
returned if the underlying `updateinfo.xml` file lacked certain
information. Pulp 2.16.3 fixes this issue.

Expand Pulp Smash's verification logic so that it checks for the
presence of erratum in content applicability reports.

See: https://pulp.plan.io/issues/3795