Replace os_is_* functions
The `pulp_smash.utils.os_is_f2{6,7}` functions are not suitable for
inclusion within a top-level toolkit. There's an enormous number of
combinations of system facts that one might want to discover, and having
`os_is_*` functions is a recipe for an explosion in number of functions.
A framework should provide more generalized tools that individual users
can adapt to their needs.

Replace these functions with functions that report information from
`/etc/os-release`. This is more suitable for use within a toolkit: there
is a much more constrained set of values that one may wish to get from
`/etc/os-release`.

These functions are still somewhat susceptible to a surface-area
explosion. If users wish to get more and more information from
`/etc/os-release`, then it would be best to create a function which
returns a dict where the keys and values correspond to the keys and
values in `/etc/os-release`. Thankfully, the new functions which were
just added are forward compatible with such a design. If these happens,
the just-added functions could be rewritten like so:

```python
def get_os_release_id(cfg, pulp_host=None):
    raise DeprecationWarning(...)
    return get_os_release(cfg, pulp_host)['ID']
```

Fix: https://github.com/PulpQE/pulp-smash/issues/1100