Lazy load the cli.Client machine
Problem:

Everytime we instantiate a new instance of `cli.Client` its `__init__` method
code tries to make communication with the remote host to determine if it is
a local or ssh machine to be used.

The communication with `machine` should be done only when it is necessary,
which is when the `run` command is called.

Solution:

Transform `machine` in a lazy `@property` then it is instantiated only
when it is first called.