Add optional timeout to settings.json
Add optional timeout variable as part of settings.json. This will allow one to
specify the timeout for the task system in seconds from now, if not
defined it will be set to default value of 1800s. This was the previous
used value.

Tweak the task system timeout to use seconds, and query Pulp more often.
Reduce the sleep time on `poll_task` function to 0.3s.

Adjust unittests, and remove old tests.

closes:#1173